refactor(infra): move setup_infra from rl/infra_setup into training.utils.infra
All loops (SFT, DPO, GRPO) share the same infra bootstrap. Delete
rl/infra_setup.py entirely; update rl/__init__.py and tests to import
directly from training.utils.infra.
