+++ training/utils/rl/__init__.py
@@ -46,7 +46,7 @@ from training.utils.rl.train import (
     run_rl_loop,
 )
 from training.utils.rl.losses import PromptGroup
-from training.utils.rl.infra_setup import Infra, setup_infra
+from training.utils.infra import Infra, setup_infra
 from training.utils.rl.metrics import (
     build_loop_metrics,
     total_target_tokens,
